+++ source/extensions/omni.isaac.orbit/test/sim/test_simulation_context.py
@@ -14,6 +14,7 @@ simulation_app = AppLauncher(headless=True).app
 
 """Rest everything follows."""
 
+import ctypes
 import numpy as np
 import traceback
 import unittest
@@ -89,12 +90,40 @@ class TestSimulationContext(unittest.TestCase):
         sim.set_setting("/myExt/using_omniverse_version", sim.get_version())
         self.assertSequenceEqual(sim.get_setting("/myExt/using_omniverse_version"), sim.get_version())
 
-    def test_render_modes(self):
-        """Test that you can change render modes."""
+    def test_headless_mode(self):
+        """Test that render mode is headless since we are running in headless mode."""
 
         sim = SimulationContext()
         # check default render mode
         self.assertEqual(sim.render_mode, sim.RenderMode.HEADLESS)
+
+    def test_boundedness(self):
+        """Test that the boundedness of the simulation context remains constant.
+
+        Note: This test fails right now because Isaac Sim does not handle boundedness correctly. On creation,
+        it is registering itself to various callbacks and hence the boundedness is more than 1. This may not be
+        critical for the simulation context since we usually call various clear functions before deleting the
+        simulation context.
+        """
+        sim = SimulationContext()
+        # manually set the boundedness to 1? -- this is not possible because of Isaac Sim.
+        sim.clear_all_callbacks()
+        sim._stage_open_callback = None
+        sim._physics_timer_callback = None
+        sim._event_timer_callback = None
+
+        # check that boundedness of simulation context is correct
+        sim_ref_count = ctypes.c_long.from_address(id(sim)).value
+        # reset the simulation
+        sim.reset()
+        self.assertEqual(ctypes.c_long.from_address(id(sim)).value, sim_ref_count)
+        # step the simulation
+        for _ in range(10):
+            sim.step()
+            self.assertEqual(ctypes.c_long.from_address(id(sim)).value, sim_ref_count)
+        # clear the simulation
+        sim.clear_instance()
+        self.assertEqual(ctypes.c_long.from_address(id(sim)).value, sim_ref_count - 1)
 
 
 if __name__ == "__main__":